bot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
autoflake==2.3.1
black==24.4.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
isort==5.13.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pyflakes==3.2.0
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.12.2
yarl==1.9.4
